Refactor web-v2 composer and scroll hooks
