+++ src/web-v2/components/conversation/ConversationComposer.tsx
@@ -1,47 +1,22 @@
-import type { ReactNode } from 'react';
-import { useCallback, useLayoutEffect, useRef, useState } from 'react';
-import { ArrowUp, Check, ChevronDown, Plus, Search } from 'lucide-react';
+import { useCallback, useRef, useState } from 'react';
+import { ArrowUp } from 'lucide-react';
 import type { ControlPlaneModelOptions } from '@web/api/client';
 import { Button } from '@web/components/ui/button';
-import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-} from '@web/components/ui/popover';
 import { Textarea } from '@web/components/ui/textarea';
 import type { ControlPlaneReasoningEffortSelection } from '@web/hooks/sessions/useControlPlaneSessionDetail';
-import type { I18nMessageKey } from '@web/i18n';
 import { useI18n } from '@web/i18n';
-import { cn } from '@web/lib/utils';
-import { FileMentionMenu } from './FileMentionMenu';
+import { useComposerTextareaAutosize } from '@web/hooks/conversation/useComposerTextareaAutosize';
+import { useFileMentionAutocomplete } from '@web/hooks/conversation/useFileMentionAutocomplete';
+import { ComposerContextMenu } from './ComposerContextMenu';
 import {
-  SessionDriftMenuSection,
-  SessionDriftStatusGlyph,
-  type SessionDriftLevel,
-} from './SessionDriftControl';
-import { useFileMentionAutocomplete } from './useFileMentionAutocomplete';
+  ComposerExecutionMenu,
+  type ComposerReasoningEffortSelection,
+} from './ComposerExecutionMenu';
+import { FileMentionMenu } from './FileMentionMenu';
+import type { SessionDriftLevel } from './SessionDriftControl';
 
-const composerTextareaMinHeight = 28;
-const composerTextareaMaxHeight = 176;
-
-type ComposerReasoningEffortSelection = Exclude<ControlPlaneReasoningEffortSelection, 'default'>;
-
-const reasoningEfforts = [
-  { value: 'low', labelKey: 'composer.reasoning.low' },
-  { value: 'medium', labelKey: 'composer.reasoning.medium' },
-  { value: 'high', labelKey: 'composer.reasoning.high' },
-  { value: 'ultrahigh', labelKey: 'composer.reasoning.ultrahigh' },
-] as const satisfies Array<{ value: ComposerReasoningEffortSelection; labelKey: I18nMessageKey }>;
-
-const driftLevelMessageKeys = {
-  unknown: 'composer.drift.signalUnknown',
-  low: 'composer.drift.signalLow',
-  medium: 'composer.drift.signalMedium',
-  high: 'composer.drift.signalHigh',
-} as const satisfies Record<SessionDriftLevel, I18nMessageKey>;
-
-// ConversationComposer owns the prompt draft and visual controls. Session
-// settings are API-backed by the parent session workflow.
+// ConversationComposer owns the prompt draft and submit lifecycle. Context,
+// execution settings, file mentions, and textarea sizing live in focused peers.
 export function ConversationComposer({
   disabled,
   driftEnabled,
@@ -62,7 +37,7 @@ export function ConversationComposer({
   driftLevel?: SessionDriftLevel;
   model?: string;
   modelOptions?: ControlPlaneModelOptions;
-  reasoningEffort?: Exclude<ControlPlaneReasoningEffortSelection, 'default'>;
+  reasoningEffort?: ComposerReasoningEffortSelection;
   settingsUpdating?: boolean;
   settingsError?: string;
   submitting?: boolean;
@@ -78,10 +53,6 @@ export function ConversationComposer({
   const effectiveDriftEnabled = driftEnabled ?? false;
   const effectiveDriftLevel = driftLevel ?? 'unknown';
   const effectiveReasoningEffort = reasoningEffort ?? 'medium';
-  const reasoningLabel = t(reasoningEfforts.find((option) => option.value === effectiveReasoningEffort)?.labelKey ?? 'composer.reasoning.medium');
-  const driftButtonLabel = effectiveDriftEnabled
-    ? `${t('composer.addContext')}: ${t(driftLevelMessageKeys[effectiveDriftLevel])}`
-    : t('composer.addContext');
 
   const handleSubmit = useCallback(async () => {
     const prompt = draft.trim();
@@ -99,18 +70,7 @@ export function ConversationComposer({
     disabled: disabled || submitting,
     onSubmit: handleSubmit,
   });
-
-  useLayoutEffect(() => {
-    const textarea = textareaRef.current;
-    if (!textarea) {
-      return;
-    }
-
-    textarea.style.height = `${composerTextareaMinHeight}px`;
-    const nextHeight = Math.min(textarea.scrollHeight, composerTextareaMaxHeight);
-    textarea.style.height = `${Math.max(nextHeight, composerTextareaMinHeight)}px`;
-    textarea.style.overflowY = textarea.scrollHeight > composerTextareaMaxHeight ? 'auto' : 'hidden';
-  }, [draft]);
+  useComposerTextareaAutosize(textareaRef, draft);
 
   return (
     <form
@@ -145,45 +105,13 @@ export function ConversationComposer({
       />
       {fileMentions.isOpen ? <FileMentionMenu {...fileMentions.menuProps} /> : null}
       <div className="v2-composer-toolbar">
-        <Popover>
-          <span className="v2-composer-context-cluster">
-            <PopoverTrigger asChild>
-              <Button
-                type="button"
-                variant="ghost"
-                size="none"
-                className="v2-composer-context-button"
-                aria-label={driftButtonLabel}
-                disabled={disabled}
-              >
-                <Plus aria-hidden="true" data-icon="inline-start" />
-              </Button>
-            </PopoverTrigger>
-            <span className="v2-composer-context-status">
-              <SessionDriftStatusGlyph
-                driftEnabled={effectiveDriftEnabled}
-                driftLevel={effectiveDriftLevel}
-              />
-            </span>
-          </span>
-          <PopoverContent
-            align="start"
-            side="top"
-            sideOffset={8}
-            className="v2-composer-menu v2-composer-context-menu"
-            aria-label={t('composer.contextMenu')}
-          >
-            {onUpdateDriftEnabled ? (
-              <SessionDriftMenuSection
-                disabled={disabled || settingsUpdating}
-                driftEnabled={effectiveDriftEnabled}
-                driftLevel={effectiveDriftLevel}
-                updating={settingsUpdating}
-                onUpdateDriftEnabled={onUpdateDriftEnabled}
-              />
-            ) : null}
-          </PopoverContent>
-        </Popover>
+        <ComposerContextMenu
+          disabled={disabled}
+          driftEnabled={effectiveDriftEnabled}
+          driftLevel={effectiveDriftLevel}
+          settingsUpdating={settingsUpdating}
+          onUpdateDriftEnabled={onUpdateDriftEnabled}
+        />
         <div className="v2-composer-toolbar-controls">
           <ComposerExecutionMenu
             disabled={disabled}
@@ -191,7 +119,6 @@ export function ConversationComposer({
             model={model}
             modelOptions={modelOptions}
             reasoningEffort={effectiveReasoningEffort}
-            reasoningLabel={reasoningLabel}
             onUpdateModel={onUpdateModel}
             onUpdateReasoningEffort={onUpdateReasoningEffort}
           />
@@ -208,197 +135,5 @@ export function ConversationComposer({
       </div>
       {settingsError ? <p className="v2-composer-error text-pretty">{settingsError}</p> : null}
     </form>
-  );
-}
-
-interface ComposerExecutionMenuProps {
-  model?: string;
-  modelOptions?: ControlPlaneModelOptions;
-  reasoningEffort: ComposerReasoningEffortSelection;
-  reasoningLabel: string;
-  disabled?: boolean;
-  settingsUpdating?: boolean;
-  onUpdateModel?: (model: string) => Promise<void>;
-  onUpdateReasoningEffort?: (value: ControlPlaneReasoningEffortSelection) => Promise<void>;
-}
-
-function ComposerExecutionMenu({
-  model,
-  modelOptions,
-  reasoningEffort,
-  reasoningLabel,
-  disabled,
-  settingsUpdating,
-  onUpdateModel,
-  onUpdateReasoningEffort,
-}: ComposerExecutionMenuProps) {
-  const { t } = useI18n();
-  const [modelSearch, setModelSearch] = useState('');
-  const groups = modelOptions?.groups ?? [];
-  const fallbackOptions = model ? [{
-    label: undefined,
-    models: [model],
-    options: [{ id: model, label: undefined, disabled: false, disabledReason: undefined }],
-  }] : [];
-  const modelGroups = groups.length ? groups : fallbackOptions;
-  const modelSearchQuery = modelSearch.trim().toLowerCase();
-  const modelOptionCount = modelGroups.reduce((count, group) => count + group.options.length, 0);
-  const showModelSearch = modelOptionCount > 6 || Boolean(modelSearchQuery);
-  const filteredModelGroups = modelSearchQuery
-    ? modelGroups
-      .map((group) => ({
-        ...group,
-        options: group.options.filter((option) => [
-          group.label,
-          option.id,
-          option.label,
-        ].some((value) => value?.toLowerCase().includes(modelSearchQuery))),
-      }))
-      .filter((group) => group.options.length)
-    : modelGroups;
-  const modelLabel = model ?? t('composer.model');
-  const triggerLabel = `${modelLabel} · ${reasoningLabel}`;
-  const triggerDisabled = disabled || settingsUpdating || (!onUpdateModel && !onUpdateReasoningEffort);
-
-  return (
-    <Popover>
-      <PopoverTrigger asChild>
-        <Button
-          type="button"
-          variant="ghost"
-          size="none"
-          className="v2-composer-execution-trigger"
-          aria-label={`${t('composer.executionMenu')}: ${triggerLabel}`}
-          aria-busy={settingsUpdating || undefined}
-          disabled={triggerDisabled}
-        >
-          <span className="v2-composer-execution-label truncate">
-            {triggerLabel}
-          </span>
-          <ChevronDown aria-hidden="true" data-icon="inline-end" />
-        </Button>
-      </PopoverTrigger>
-      <PopoverContent
-        align="end"
-        side="top"
-        sideOffset={8}
-        className="v2-composer-menu v2-composer-execution-menu"
-        aria-label={t('composer.executionMenu')}
-      >
-        <div className="v2-composer-menu-section">
-          <p className="v2-composer-menu-heading">
-            {t('composer.reasoningEffort')}
-          </p>
-          <div className="v2-composer-menu-options">
-            {reasoningEfforts.map((option) => (
-              <ComposerMenuOption
-                key={option.value}
-                compact
-                selected={reasoningEffort === option.value}
-                disabled={!onUpdateReasoningEffort || settingsUpdating}
-                onSelect={() => {
-                  void onUpdateReasoningEffort?.(option.value);
-                }}
-              >
-                {t(option.labelKey)}
-              </ComposerMenuOption>
-            ))}
-          </div>
-        </div>
-        <div className="v2-composer-menu-section">
-          <p className="v2-composer-menu-heading">
-            {t('composer.model')}
-          </p>
-          {showModelSearch ? (
-            <label className="v2-composer-model-search">
-              <Search aria-hidden="true" data-icon="inline-start" />
-              <input
-                type="text"
-                className="v2-composer-model-search-input"
-                value={modelSearch}
-                placeholder={t('composer.searchModels')}
-                aria-label={t('composer.searchModels')}
-                autoComplete="off"
-                autoCorrect="off"
-                spellCheck={false}
-                onChange={(event) => {
-                  setModelSearch(event.target.value);
-                }}
-              />
-            </label>
-          ) : null}
-          <div className="v2-composer-menu-options">
-            {filteredModelGroups.map((group) => (
-              <div key={group.label ?? group.models.join(',')} className="v2-composer-menu-option-group">
-                {group.label ? (
-                  <p className="v2-composer-menu-group-heading">
-                    {group.label}
-                  </p>
-                ) : null}
-                {group.options.map((option) => (
-                  <ComposerMenuOption
-                    key={option.id}
-                    selected={model === option.id}
-                    disabled={!onUpdateModel || settingsUpdating || option.disabled}
-                    description={option.disabledReason}
-                    onSelect={() => {
-                      void onUpdateModel?.(option.id);
-                    }}
-                  >
-                    {option.label ?? option.id}
-                  </ComposerMenuOption>
-                ))}
-              </div>
-            ))}
-            {filteredModelGroups.length ? null : (
-              <p className="v2-composer-menu-empty text-pretty">
-                {t('composer.noModelMatches')}
-              </p>
-            )}
-          </div>
-        </div>
-      </PopoverContent>
-    </Popover>
-  );
-}
-
-function ComposerMenuOption({
-  children,
-  description,
-  compact,
-  selected,
-  disabled,
-  onSelect,
-}: {
-  children: ReactNode;
-  description?: string;
-  compact?: boolean;
-  selected?: boolean;
-  disabled?: boolean;
-  onSelect: () => void;
-}) {
-  return (
-    <Button
-      type="button"
-      variant="ghost"
-      size="none"
-      className={cn('v2-composer-menu-option justify-between text-left', compact && 'v2-composer-menu-option-compact')}
-      role="menuitemradio"
-      aria-checked={selected}
-      disabled={disabled}
-      onClick={onSelect}
-    >
-      <span className="v2-composer-menu-option-copy">
-        <span className="v2-composer-menu-option-label truncate">
-          {children}
-        </span>
-        {description ? (
-          <span className="v2-composer-menu-option-description truncate">
-            {description}
-          </span>
-        ) : null}
-      </span>
-      {selected ? <Check aria-hidden="true" data-icon="inline-end" /> : null}
-    </Button>
   );
 }
