+++ src/web-v2/components/conversation/ConversationThread.tsx
@@ -1,4 +1,3 @@
-import { useCallback, useEffect, useLayoutEffect, useMemo, useRef, type KeyboardEvent } from 'react';
 import type { ControlPlaneModelOptions } from '@web/api/client';
 import type {
   ControlPlaneApprovalDecision,
@@ -6,20 +5,11 @@ import type {
   ControlPlanePendingApproval,
   ControlPlaneSessionDetail,
 } from '@web/hooks/sessions/useControlPlaneSessionDetail';
+import { useConversationAutoScroll } from '@web/hooks/conversation/useConversationAutoScroll';
 import { ApprovalPanel } from './ApprovalPanel';
 import { ConversationComposer } from './ConversationComposer';
 import { ConversationMessage } from './ConversationMessage';
 import { Loader2 } from 'lucide-react';
-
-const userScrollKeys = new Set([
-  'ArrowDown',
-  'ArrowUp',
-  'End',
-  'Home',
-  'PageDown',
-  'PageUp',
-  ' ',
-]);
 
 interface ConversationThreadProps {
   session: ControlPlaneSessionDetail;
@@ -59,106 +49,12 @@ export function ConversationThread({
   onUpdateReasoningEffort,
   onResolveApproval,
 }: ConversationThreadProps) {
-  const scrollContainerRef = useRef<HTMLDivElement>(null);
-  const autoScrollEnabledRef = useRef(true);
-  const userScrollIntentRef = useRef(false);
-  const previousResponseActiveRef = useRef(false);
-  const programmaticScrollFrameRef = useRef<number | undefined>(undefined);
-  const responseActive = Boolean(
-    submitting ||
-    liveStatus ||
-    session?.messages.some((message) => message.isStreaming || message.isPending),
-  );
-  const conversationScrollKey = useMemo(() => {
-    if (!session) {
-      return '';
-    }
-
-    const lastMessage = session.messages.at(-1);
-    return [
-      session.id,
-      session.messages.length,
-      lastMessage?.id,
-      lastMessage?.text.length,
-      lastMessage?.isStreaming,
-      lastMessage?.isPending,
-      liveStatus,
-    ].join(':');
-  }, [liveStatus, session]);
-  const scrollToBottom = useCallback(() => {
-    const container = scrollContainerRef.current;
-    if (!container) {
-      return;
-    }
-
-    if (programmaticScrollFrameRef.current) {
-      window.cancelAnimationFrame(programmaticScrollFrameRef.current);
-    }
-
-    userScrollIntentRef.current = false;
-    container.dataset.programmaticScroll = 'true';
-    container.scrollTop = container.scrollHeight;
-    programmaticScrollFrameRef.current = window.requestAnimationFrame(() => {
-      delete container.dataset.programmaticScroll;
-      programmaticScrollFrameRef.current = undefined;
-    });
-  }, []);
-  const handleUserScrollIntent = useCallback(() => {
-    if (responseActive) {
-      userScrollIntentRef.current = true;
-    }
-  }, [responseActive]);
-  const handleScrollKeyDown = useCallback((event: KeyboardEvent<HTMLDivElement>) => {
-    if (userScrollKeys.has(event.key)) {
-      handleUserScrollIntent();
-    }
-  }, [handleUserScrollIntent]);
-  const handleConversationScroll = useCallback(() => {
-    const container = scrollContainerRef.current;
-    if (!container || container.dataset.programmaticScroll === 'true') {
-      return;
-    }
-
-    if (responseActive && userScrollIntentRef.current) {
-      autoScrollEnabledRef.current = false;
-    }
-  }, [responseActive]);
-
-  useEffect(() => () => {
-    if (programmaticScrollFrameRef.current) {
-      window.cancelAnimationFrame(programmaticScrollFrameRef.current);
-    }
-  }, []);
-
-  useLayoutEffect(() => {
-    if (!session?.id) {
-      return;
-    }
-
-    autoScrollEnabledRef.current = true;
-    userScrollIntentRef.current = false;
-    scrollToBottom();
-  }, [scrollToBottom, session?.id]);
-
-  useEffect(() => {
-    if (responseActive && !previousResponseActiveRef.current) {
-      autoScrollEnabledRef.current = true;
-      userScrollIntentRef.current = false;
-      scrollToBottom();
-    }
-
-    if (!responseActive) {
-      userScrollIntentRef.current = false;
-    }
-
-    previousResponseActiveRef.current = responseActive;
-  }, [responseActive, scrollToBottom]);
-
-  useLayoutEffect(() => {
-    if (autoScrollEnabledRef.current) {
-      scrollToBottom();
-    }
-  }, [conversationScrollKey, scrollToBottom]);
+  const conversationAutoScroll = useConversationAutoScroll({
+    liveStatus,
+    messages: session?.messages ?? [],
+    sessionId: session?.id,
+    submitting,
+  });
 
   if (loading && !session) {
     return (
@@ -192,13 +88,9 @@ export function ConversationThread({
   return (
     <div className="flex h-full min-w-0 flex-col">
       <div
-        ref={scrollContainerRef}
+        ref={conversationAutoScroll.scrollContainerRef}
         className="v2-conversation-scroll v2-scrollbar-hidden min-h-0 flex-1 overflow-auto"
-        onKeyDown={handleScrollKeyDown}
-        onPointerDown={handleUserScrollIntent}
-        onScroll={handleConversationScroll}
-        onTouchStart={handleUserScrollIntent}
-        onWheel={handleUserScrollIntent}
+        {...conversationAutoScroll.scrollContainerProps}
       >
         <div className="mx-auto flex w-full max-w-3xl flex-col gap-5 px-6 py-8">
           {loading ? (
